src/model_improvement_toolkit/__init__.py
"""Model Improvement Toolkit

機械学習モデル（LightGBMとXGBoost）の改善のための示唆を提供するPythonパッケージ
"""

__version__ = "0.1.0"
__author__ = "Model Improvement Toolkit Contributors"
__email__ = "contact@example.com"

from typing import Any, Dict, Optional, Union
import pandas as pd

def analyze_model(
    model: Any,
    X_train: pd.DataFrame,
    y_train: pd.Series,
    X_val: Optional[pd.DataFrame] = None,
    y_val: Optional[pd.Series] = None,
    model_type: Optional[str] = None
) -> Dict[str, Any]:
    """
    モデルの包括的な分析を実行
    
    Parameters
    ----------
    model : Any
        学習済みのLightGBMまたはXGBoostモデル
    X_train : pd.DataFrame
        訓練データの特徴量
    y_train : pd.Series
        訓練データのターゲット
    X_val : pd.DataFrame, optional
        検証データの特徴量
    y_val : pd.Series, optional
        検証データのターゲット
    model_type : str, optional
        モデルタイプ ("lightgbm" or "xgboost")。指定しない場合は自動検出
    
    Returns
    -------
    Dict[str, Any]
        分析結果を含む辞書
    """
    # TODO: 実装
    raise NotImplementedError("This function will be implemented in Phase 2")

__all__ = ["analyze_model", "__version__"]