requirements-dev.txt
# Include base requirements
-r requirements.txt

# Testing
pytest>=7.0
pytest-cov>=4.0
pytest-mock>=3.10

# Code quality
black>=23.0
mypy>=1.0
flake8>=6.0
isort>=5.0
bandit>=1.7

# Pre-commit hooks
pre-commit>=3.0

# Documentation
sphinx>=5.0
sphinx-rtd-theme>=1.0
nbsphinx>=0.8

# Type stubs
types-requests
pandas-stubs